Clean pull request issues
