+++ BestRecipes/Aplication/SceneDelegate.swift
@@ -14,7 +14,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let scene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: scene)
-      window?.rootViewController = DetailViewController(detailView: DetailView())
+        window?.rootViewController = DetailViewController(detailView: DetailView())
         window?.makeKeyAndVisible()
     }
